test(slack): Split conversation work component coverage
Move Slack-specific durable mailbox coverage out of the generic conversation work component test and share local fixtures across the split files.

This keeps each component test file below the large-file threshold without changing the covered behavior.
