+++ packages/junior/tests/component/task-execution/conversation-work.test.ts
@@ -1,7 +1,4 @@
-import { createHmac } from "node:crypto";
 import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
-import type { Message, StateAdapter, Thread } from "chat";
-import type { ConversationWorkQueue } from "@/chat/task-execution/queue";
 import { recoverConversationWork } from "@/chat/task-execution/heartbeat";
 import { runHeartbeat } from "@/chat/agent-dispatch/heartbeat";
 import {
@@ -22,190 +19,17 @@ import {
   CONVERSATION_WORK_DEFER_DELAY_MS,
   processConversationWork,
 } from "@/chat/task-execution/worker";
-import { handleSlackWebhook } from "@/chat/ingress/slack-webhook";
-import { createSlackConversationWorker } from "@/chat/task-execution/slack-work";
 import { processConversationQueueMessage } from "@/chat/task-execution/vercel-callback";
 import { createVercelConversationWorkQueue } from "@/chat/task-execution/vercel-queue";
-import { createJuniorSlackAdapter } from "@/chat/slack/adapter";
 import { disconnectStateAdapter, getStateAdapter } from "@/chat/state/adapter";
-import { upsertAgentTurnSessionRecord } from "@/chat/state/turn-session";
-import type { WaitUntilFn } from "@/handlers/types";
-
-vi.hoisted(() => {
-  process.env.JUNIOR_STATE_ADAPTER = "memory";
-});
-
-const CONVERSATION_ID = "slack:C123:1712345.0001";
-const SLACK_SIGNING_SECRET = "test-signing-secret";
-const SLACK_BOT_USER_ID = "U_BOT";
-
-class FakeQueue implements ConversationWorkQueue {
-  fail = false;
-  sent: Array<{
-    conversationId: string;
-    delayMs?: number;
-    idempotencyKey?: string;
-  }> = [];
-
-  async send(
-    message: { conversationId: string },
-    options?: { delayMs?: number; idempotencyKey?: string },
-  ): Promise<{ messageId: string }> {
-    if (this.fail) {
-      throw new Error("queue unavailable");
-    }
-    this.sent.push({
-      conversationId: message.conversationId,
-      delayMs: options?.delayMs,
-      idempotencyKey: options?.idempotencyKey,
-    });
-    return { messageId: `queue-${this.sent.length}` };
-  }
-}
-
-function inboundMessage(
-  inboundMessageId: string,
-  overrides: Partial<InboundMessageRecord> = {},
-): InboundMessageRecord {
-  return {
-    conversationId: CONVERSATION_ID,
-    inboundMessageId,
-    source: "slack",
-    createdAtMs: 1_000,
-    receivedAtMs: 1_100,
-    input: {
-      text: `message ${inboundMessageId}`,
-      authorId: "U123",
-    },
-    ...overrides,
-  };
-}
-
-function delayIndexLockOnce(state: StateAdapter): StateAdapter {
-  let blocked = false;
-  return new Proxy(state, {
-    get(target, prop, receiver) {
-      if (prop === "acquireLock") {
-        return async (key: string, ttlMs: number) => {
-          if (!blocked && key === "junior:conversation-work:index:lock") {
-            blocked = true;
-            return null;
-          }
-          return target.acquireLock(key, ttlMs);
-        };
-      }
-      const value = Reflect.get(target, prop, receiver);
-      return typeof value === "function" ? value.bind(target) : value;
-    },
-  }) as StateAdapter;
-}
-
-function delayMutationLockUntil(args: {
-  conversationId: string;
-  readyAtMs: number;
-  state: StateAdapter;
-}): StateAdapter {
-  const mutationLockKey = `junior:conversation-work:mutation:${args.conversationId}`;
-  return new Proxy(args.state, {
-    get(target, prop, receiver) {
-      if (prop === "acquireLock") {
-        return async (key: string, ttlMs: number) => {
-          if (key === mutationLockKey && Date.now() < args.readyAtMs) {
-            return null;
-          }
-          return target.acquireLock(key, ttlMs);
-        };
-      }
-      const value = Reflect.get(target, prop, receiver);
-      return typeof value === "function" ? value.bind(target) : value;
-    },
-  }) as StateAdapter;
-}
-
-function signSlackBody(body: string, timestamp: string): string {
-  return `v0=${createHmac("sha256", SLACK_SIGNING_SECRET)
-    .update(`v0:${timestamp}:${body}`)
-    .digest("hex")}`;
-}
-
-function slackWebhookRequest(body: unknown): Request {
-  const serialized = JSON.stringify(body);
-  const timestamp = String(Math.floor(Date.now() / 1000));
-  return new Request("https://example.test/api/webhooks/slack", {
-    method: "POST",
-    headers: {
-      "content-type": "application/json",
-      "x-slack-request-timestamp": timestamp,
-      "x-slack-signature": signSlackBody(serialized, timestamp),
-    },
-    body: serialized,
-  });
-}
-
-function slackEnvelope(input: {
-  channel?: string;
-  eventType?: "app_mention" | "message";
-  text?: string;
-  threadTs?: string;
-  ts?: string;
-}) {
-  const channel = input.channel ?? "C123";
-  const ts = input.ts ?? "1712345.0001";
-  return {
-    team_id: "T123",
-    type: "event_callback",
-    event: {
-      type: input.eventType ?? "app_mention",
-      user: "U123",
-      text: input.text ?? `<@${SLACK_BOT_USER_ID}> hello`,
-      channel,
-      ts,
-      event_ts: ts,
-      channel_type: channel.startsWith("D") ? "im" : "channel",
-      ...(input.threadTs ? { thread_ts: input.threadTs } : {}),
-    },
-  };
-}
-
-function deferred<T>(): {
-  promise: Promise<T>;
-  reject(error: unknown): void;
-  resolve(value: T): void;
-} {
-  let resolve!: (value: T) => void;
-  let reject!: (error: unknown) => void;
-  const promise = new Promise<T>((resolvePromise, rejectPromise) => {
-    resolve = resolvePromise;
-    reject = rejectPromise;
-  });
-  return { promise, resolve, reject };
-}
-
-type WaitUntilTask = () => Promise<unknown>;
-
-function collectWaitUntil(tasks: WaitUntilTask[]): WaitUntilFn {
-  return (task) => {
-    tasks.push(typeof task === "function" ? task : () => task);
-  };
-}
-
-async function flushWaitUntil(tasks: WaitUntilTask[]): Promise<void> {
-  for (let index = 0; index < tasks.length; index += 1) {
-    await tasks[index]?.();
-  }
-}
-
-async function handleSlackWebhookAndFlush(
-  args: Omit<Parameters<typeof handleSlackWebhook>[0], "waitUntil">,
-): Promise<Response> {
-  const waitUntilTasks: WaitUntilTask[] = [];
-  const response = await handleSlackWebhook({
-    ...args,
-    waitUntil: collectWaitUntil(waitUntilTasks),
-  });
-  await flushWaitUntil(waitUntilTasks);
-  return response;
-}
+import {
+  CONVERSATION_ID,
+  FakeQueue,
+  deferred,
+  delayIndexLockOnce,
+  delayMutationLockUntil,
+  inboundMessage,
+} from "./conversation-work-fixtures";
 
 describe("conversation work execution", () => {
   beforeEach(async () => {
@@ -787,660 +611,6 @@ describe("conversation work execution", () => {
     ).resolves.toEqual({ status: "completed" });
 
     expect(injected).toEqual(["m1"]);
-  });
-
-  it("persists Slack mentions into the durable mailbox and wakes the queue", async () => {
-    const queue = new FakeQueue();
-    const state = getStateAdapter();
-    await state.connect();
-    const slackAdapter = createJuniorSlackAdapter({
-      botToken: "xoxb-test",
-      botUserId: SLACK_BOT_USER_ID,
-      signingSecret: SLACK_SIGNING_SECRET,
-    });
-
-    const response = await handleSlackWebhookAndFlush({
-      request: slackWebhookRequest(
-        slackEnvelope({
-          text: `<@${SLACK_BOT_USER_ID}> deploy status`,
-        }),
-      ),
-      services: {
-        getSlackAdapter: () => slackAdapter,
-        queue,
-        runtime: {
-          handleAssistantContextChanged: async () => {},
-          handleAssistantThreadStarted: async () => {},
-          handleNewMention: async () => {},
-          handleSubscribedMessage: async () => {},
-        },
-        state,
-      },
-    });
-
-    expect(response.status).toBe(200);
-    expect(queue.sent).toEqual([
-      expect.objectContaining({
-        conversationId: CONVERSATION_ID,
-      }),
-    ]);
-    const work = await getConversationWorkState({
-      conversationId: CONVERSATION_ID,
-      state,
-    });
-    expect(work?.needsRun).toBe(true);
-    expect(work?.messages).toEqual([
-      expect.objectContaining({
-        conversationId: CONVERSATION_ID,
-        source: "slack",
-        input: expect.objectContaining({
-          authorId: "U123",
-          metadata: expect.objectContaining({
-            platform: "slack",
-            route: "mention",
-          }),
-        }),
-      }),
-    ]);
-  });
-
-  it("runs queued Slack mailbox work through the Slack runtime", async () => {
-    const queue = new FakeQueue();
-    const state = getStateAdapter();
-    await state.connect();
-    const slackAdapter = createJuniorSlackAdapter({
-      botToken: "xoxb-test",
-      botUserId: SLACK_BOT_USER_ID,
-      signingSecret: SLACK_SIGNING_SECRET,
-    });
-    const calls: Array<{
-      message: Message;
-      skipped: Message[];
-      thread: Thread;
-    }> = [];
-
-    await handleSlackWebhookAndFlush({
-      request: slackWebhookRequest(
-        slackEnvelope({
-          text: `<@${SLACK_BOT_USER_ID}> first`,
-          ts: "1712345.0001",
-        }),
-      ),
-      services: {
-        getSlackAdapter: () => slackAdapter,
-        queue,
-        runtime: {
-          handleAssistantContextChanged: async () => {},
-          handleAssistantThreadStarted: async () => {},
-          handleNewMention: async () => {},
-          handleSubscribedMessage: async () => {},
-        },
-        state,
-      },
-    });
-    await handleSlackWebhookAndFlush({
-      request: slackWebhookRequest(
-        slackEnvelope({
-          text: `<@${SLACK_BOT_USER_ID}> second`,
-          ts: "1712345.0002",
-          threadTs: "1712345.0001",
-        }),
-      ),
-      services: {
-        getSlackAdapter: () => slackAdapter,
-        queue,
-        runtime: {
-          handleAssistantContextChanged: async () => {},
-          handleAssistantThreadStarted: async () => {},
-          handleNewMention: async () => {},
-          handleSubscribedMessage: async () => {},
-        },
-        state,
-      },
-    });
-
-    await expect(
-      processConversationWork(CONVERSATION_ID, {
-        queue,
-        state,
-        run: createSlackConversationWorker({
-          getSlackAdapter: () => slackAdapter,
-          runtime: {
-            handleNewMention: async (thread, message, hooks) => {
-              calls.push({
-                thread,
-                message,
-                skipped: hooks?.messageContext?.skipped ?? [],
-              });
-            },
-            handleSubscribedMessage: async () => {
-              throw new Error("unexpected subscribed route");
-            },
-          },
-          state,
-        }),
-      }),
-    ).resolves.toEqual({ status: "completed" });
-
-    expect(calls).toHaveLength(1);
-    expect(calls[0]?.thread.id).toBe(CONVERSATION_ID);
-    expect(calls[0]?.message.id).toBe("1712345.0002");
-    expect(calls[0]?.message.text).toContain("second");
-    expect(calls[0]?.skipped.map((message) => message.id)).toEqual([
-      "1712345.0001",
-    ]);
-    const work = await getConversationWorkState({
-      conversationId: CONVERSATION_ID,
-      state,
-    });
-    expect(work ? countPendingConversationMessages(work) : 0).toBe(0);
-  });
-
-  it("keeps restored thread context aligned with promoted mention routing", async () => {
-    const queue = new FakeQueue();
-    const state = getStateAdapter();
-    await state.connect();
-    const slackAdapter = createJuniorSlackAdapter({
-      botToken: "xoxb-test",
-      botUserId: SLACK_BOT_USER_ID,
-      signingSecret: SLACK_SIGNING_SECRET,
-    });
-    const calls: Array<{
-      message: Message;
-      skipped: Message[];
-      thread: Thread;
-    }> = [];
-    const subscribedValues: boolean[] = [];
-    const ingressServices = {
-      getSlackAdapter: () => slackAdapter,
-      queue,
-      runtime: {
-        handleAssistantContextChanged: async () => {},
-        handleAssistantThreadStarted: async () => {},
-        handleNewMention: async () => {},
-        handleSubscribedMessage: async () => {},
-      },
-      state,
-    };
-
-    await handleSlackWebhookAndFlush({
-      request: slackWebhookRequest(
-        slackEnvelope({
-          text: `<@${SLACK_BOT_USER_ID}> first`,
-          ts: "1712345.0001",
-        }),
-      ),
-      services: ingressServices,
-    });
-    await state.subscribe(CONVERSATION_ID);
-    await handleSlackWebhookAndFlush({
-      request: slackWebhookRequest(
-        slackEnvelope({
-          eventType: "message",
-          text: "follow-up without an explicit mention",
-          ts: "1712345.0002",
-          threadTs: "1712345.0001",
-        }),
-      ),
-      services: ingressServices,
-    });
-    const workBeforeProcessing = await getConversationWorkState({
-      conversationId: CONVERSATION_ID,
-      state,
-    });
-    expect(
-      workBeforeProcessing?.messages.map((record) => record.input.metadata),
-    ).toEqual([
-      expect.objectContaining({ route: "mention" }),
-      expect.objectContaining({ route: "subscribed" }),
-    ]);
-    await state.unsubscribe(CONVERSATION_ID);
-
-    await expect(
-      processConversationWork(CONVERSATION_ID, {
-        queue,
-        state,
-        run: createSlackConversationWorker({
-          getSlackAdapter: () => slackAdapter,
-          runtime: {
-            handleNewMention: async (thread, message, hooks) => {
-              subscribedValues.push(await thread.isSubscribed());
-              calls.push({
-                thread,
-                message,
-                skipped: hooks?.messageContext?.skipped ?? [],
-              });
-            },
-            handleSubscribedMessage: async () => {
-              throw new Error(
-                "mixed mention batches should promote to mention",
-              );
-            },
-          },
-          state,
-        }),
-      }),
-    ).resolves.toEqual({ status: "completed" });
-
-    expect(calls).toHaveLength(1);
-    expect(calls[0]?.message.id).toBe("1712345.0002");
-    expect(calls[0]?.skipped.map((message) => message.id)).toEqual([
-      "1712345.0001",
-    ]);
-    expect(subscribedValues).toEqual([false]);
-  });
-
-  it("processes pending Slack follow-ups before timeout continuation", async () => {
-    const queue = new FakeQueue();
-    const state = getStateAdapter();
-    await state.connect();
-    const slackAdapter = createJuniorSlackAdapter({
-      botToken: "xoxb-test",
-      botUserId: SLACK_BOT_USER_ID,
-      signingSecret: SLACK_SIGNING_SECRET,
-    });
-    await upsertAgentTurnSessionRecord({
-      conversationId: CONVERSATION_ID,
-      sessionId: "turn-timeout",
-      sliceId: 2,
-      state: "awaiting_resume",
-      resumeReason: "timeout",
-      piMessages: [
-        {
-          role: "user",
-          content: [{ type: "text", text: "original request" }],
-          timestamp: 1_000,
-        },
-      ],
-    });
-
-    await handleSlackWebhookAndFlush({
-      request: slackWebhookRequest(
-        slackEnvelope({
-          text: `<@${SLACK_BOT_USER_ID}> follow-up`,
-          ts: "1712345.0002",
-          threadTs: "1712345.0001",
-        }),
-      ),
-      services: {
-        getSlackAdapter: () => slackAdapter,
-        queue,
-        runtime: {
-          handleAssistantContextChanged: async () => {},
-          handleAssistantThreadStarted: async () => {},
-          handleNewMention: async () => {},
-          handleSubscribedMessage: async () => {},
-        },
-        state,
-      },
-    });
-
-    const calls: string[] = [];
-    await expect(
-      processConversationWork(CONVERSATION_ID, {
-        queue,
-        state,
-        run: createSlackConversationWorker({
-          getSlackAdapter: () => slackAdapter,
-          runtime: {
-            handleNewMention: async (_thread, message) => {
-              calls.push(message.text);
-            },
-            handleSubscribedMessage: async () => {
-              throw new Error("unexpected subscribed route");
-            },
-          },
-          state,
-        }),
-      }),
-    ).resolves.toEqual({ status: "completed" });
-
-    expect(calls).toEqual([expect.stringContaining("follow-up")]);
-    const work = await getConversationWorkState({
-      conversationId: CONVERSATION_ID,
-      state,
-    });
-    expect(work ? countPendingConversationMessages(work) : 0).toBe(0);
-  });
-
-  it("drains Slack messages that arrive during an active turn into steering", async () => {
-    const queue = new FakeQueue();
-    const state = getStateAdapter();
-    await state.connect();
-    const slackAdapter = createJuniorSlackAdapter({
-      botToken: "xoxb-test",
-      botUserId: SLACK_BOT_USER_ID,
-      signingSecret: SLACK_SIGNING_SECRET,
-    });
-    const ingressServices = {
-      getSlackAdapter: () => slackAdapter,
-      queue,
-      runtime: {
-        handleAssistantContextChanged: async () => {},
-        handleAssistantThreadStarted: async () => {},
-        handleNewMention: async () => {},
-        handleSubscribedMessage: async () => {},
-      },
-      state,
-    };
-    await handleSlackWebhookAndFlush({
-      request: slackWebhookRequest(
-        slackEnvelope({
-          text: `<@${SLACK_BOT_USER_ID}> first`,
-          ts: "1712345.0001",
-        }),
-      ),
-      services: ingressServices,
-    });
-
-    const injected: string[][] = [];
-    const drained: string[][] = [];
-    await expect(
-      processConversationWork(CONVERSATION_ID, {
-        queue,
-        state,
-        run: createSlackConversationWorker({
-          getSlackAdapter: () => slackAdapter,
-          runtime: {
-            handleNewMention: async (_thread, _message, hooks) => {
-              await hooks?.onTurnStatePersisted?.();
-              await handleSlackWebhookAndFlush({
-                request: slackWebhookRequest(
-                  slackEnvelope({
-                    text: `<@${SLACK_BOT_USER_ID}> steer this`,
-                    ts: "1712345.0002",
-                    threadTs: "1712345.0001",
-                  }),
-                ),
-                services: ingressServices,
-              });
-              const messages =
-                (await hooks?.drainSteeringMessages?.(async (steering) => {
-                  injected.push(steering.map((message) => message.id));
-                })) ?? [];
-              drained.push(messages.map((message) => message.id));
-            },
-            handleSubscribedMessage: async () => {
-              throw new Error("unexpected subscribed route");
-            },
-          },
-          state,
-        }),
-      }),
-    ).resolves.toEqual({ status: "completed" });
-
-    expect(injected).toEqual([["1712345.0002"]]);
-    expect(drained).toEqual([["1712345.0002"]]);
-    const work = await getConversationWorkState({
-      conversationId: CONVERSATION_ID,
-      state,
-    });
-    expect(work?.messages.map((message) => message.injectedAtMs)).toEqual([
-      expect.any(Number),
-      expect.any(Number),
-    ]);
-    expect(work ? countPendingConversationMessages(work) : 0).toBe(0);
-  });
-
-  it("does not replay injected Slack mailbox records after lease recovery", async () => {
-    const queue = new FakeQueue();
-    const state = getStateAdapter();
-    await state.connect();
-    const slackAdapter = createJuniorSlackAdapter({
-      botToken: "xoxb-test",
-      botUserId: SLACK_BOT_USER_ID,
-      signingSecret: SLACK_SIGNING_SECRET,
-    });
-
-    await handleSlackWebhookAndFlush({
-      request: slackWebhookRequest(
-        slackEnvelope({
-          text: `<@${SLACK_BOT_USER_ID}> first`,
-        }),
-      ),
-      services: {
-        getSlackAdapter: () => slackAdapter,
-        queue,
-        runtime: {
-          handleAssistantContextChanged: async () => {},
-          handleAssistantThreadStarted: async () => {},
-          handleNewMention: async () => {},
-          handleSubscribedMessage: async () => {},
-        },
-        state,
-      },
-    });
-    const lease = await startConversationWork({
-      conversationId: CONVERSATION_ID,
-      nowMs: 2_000,
-      state,
-    });
-    expect(lease.status).toBe("acquired");
-    if (lease.status !== "acquired") {
-      return;
-    }
-    const work = await getConversationWorkState({
-      conversationId: CONVERSATION_ID,
-      state,
-    });
-    const inboundMessageIds =
-      work?.messages.map((message) => message.inboundMessageId) ?? [];
-    await markConversationMessagesInjected({
-      conversationId: CONVERSATION_ID,
-      inboundMessageIds,
-      leaseToken: lease.leaseToken,
-      nowMs: 3_000,
-      state,
-    });
-    await recoverConversationWork({
-      nowMs: 2_000 + CONVERSATION_WORK_LEASE_TTL_MS,
-      queue,
-      state,
-    });
-
-    await expect(
-      processConversationWork(CONVERSATION_ID, {
-        queue,
-        state,
-        run: createSlackConversationWorker({
-          getSlackAdapter: () => slackAdapter,
-          runtime: {
-            handleNewMention: async () => {
-              throw new Error("injected messages should not replay");
-            },
-            handleSubscribedMessage: async () => {
-              throw new Error("injected messages should not replay");
-            },
-          },
-          state,
-        }),
-      }),
-    ).resolves.toEqual({ status: "no_work" });
-
-    const recovered = await getConversationWorkState({
-      conversationId: CONVERSATION_ID,
-      state,
-    });
-    expect(recovered?.needsRun).toBe(false);
-    expect(recovered ? countPendingConversationMessages(recovered) : 0).toBe(0);
-  });
-
-  it("keeps Slack mailbox records pending when the runtime handoff fails", async () => {
-    const queue = new FakeQueue();
-    const state = getStateAdapter();
-    await state.connect();
-    const slackAdapter = createJuniorSlackAdapter({
-      botToken: "xoxb-test",
-      botUserId: SLACK_BOT_USER_ID,
-      signingSecret: SLACK_SIGNING_SECRET,
-    });
-
-    await handleSlackWebhookAndFlush({
-      request: slackWebhookRequest(
-        slackEnvelope({
-          text: `<@${SLACK_BOT_USER_ID}> first`,
-        }),
-      ),
-      services: {
-        getSlackAdapter: () => slackAdapter,
-        queue,
-        runtime: {
-          handleAssistantContextChanged: async () => {},
-          handleAssistantThreadStarted: async () => {},
-          handleNewMention: async () => {},
-          handleSubscribedMessage: async () => {},
-        },
-        state,
-      },
-    });
-
-    await expect(
-      processConversationWork(CONVERSATION_ID, {
-        queue,
-        state,
-        run: createSlackConversationWorker({
-          getSlackAdapter: () => slackAdapter,
-          runtime: {
-            handleNewMention: async () => {
-              throw new Error("runtime failed before durable handoff");
-            },
-            handleSubscribedMessage: async () => {
-              throw new Error("unexpected subscribed route");
-            },
-          },
-          state,
-        }),
-      }),
-    ).rejects.toThrow("runtime failed before durable handoff");
-
-    const work = await getConversationWorkState({
-      conversationId: CONVERSATION_ID,
-      state,
-    });
-    expect(work?.lease).toBeUndefined();
-    expect(work ? countPendingConversationMessages(work) : 0).toBe(1);
-    expect(work?.messages[0]?.injectedAtMs).toBeUndefined();
-  });
-
-  it("reports lost lease when Slack injection marking loses ownership", async () => {
-    const queue = new FakeQueue();
-    const state = getStateAdapter();
-    await state.connect();
-    const slackAdapter = createJuniorSlackAdapter({
-      botToken: "xoxb-test",
-      botUserId: SLACK_BOT_USER_ID,
-      signingSecret: SLACK_SIGNING_SECRET,
-    });
-
-    await handleSlackWebhookAndFlush({
-      request: slackWebhookRequest(
-        slackEnvelope({
-          text: `<@${SLACK_BOT_USER_ID}> first`,
-        }),
-      ),
-      services: {
-        getSlackAdapter: () => slackAdapter,
-        queue,
-        runtime: {
-          handleAssistantContextChanged: async () => {},
-          handleAssistantThreadStarted: async () => {},
-          handleNewMention: async () => {},
-          handleSubscribedMessage: async () => {},
-        },
-        state,
-      },
-    });
-
-    let handled = 0;
-    const worker = createSlackConversationWorker({
-      getSlackAdapter: () => slackAdapter,
-      runtime: {
-        handleNewMention: async () => {
-          handled += 1;
-        },
-        handleSubscribedMessage: async () => {
-          throw new Error("unexpected subscribed route");
-        },
-      },
-      state,
-    });
-
-    await expect(
-      worker({
-        checkIn: async () => true,
-        conversationId: CONVERSATION_ID,
-        drainMailbox: async () => [],
-        leaseToken: "stale-lease",
-        shouldYield: () => false,
-      }),
-    ).resolves.toEqual({ status: "lost_lease" });
-
-    expect(handled).toBe(1);
-    const work = await getConversationWorkState({
-      conversationId: CONVERSATION_ID,
-      state,
-    });
-    expect(work ? countPendingConversationMessages(work) : 0).toBe(1);
-  });
-
-  it("completes Slack mailbox work when the handler finishes after the soft deadline", async () => {
-    const queue = new FakeQueue();
-    const state = getStateAdapter();
-    await state.connect();
-    const slackAdapter = createJuniorSlackAdapter({
-      botToken: "xoxb-test",
-      botUserId: SLACK_BOT_USER_ID,
-      signingSecret: SLACK_SIGNING_SECRET,
-    });
-    let currentNowMs = 1_000;
-
-    await handleSlackWebhookAndFlush({
-      request: slackWebhookRequest(
-        slackEnvelope({
-          text: `<@${SLACK_BOT_USER_ID}> first`,
-        }),
-      ),
-      services: {
-        getSlackAdapter: () => slackAdapter,
-        queue,
-        runtime: {
-          handleAssistantContextChanged: async () => {},
-          handleAssistantThreadStarted: async () => {},
-          handleNewMention: async () => {},
-          handleSubscribedMessage: async () => {},
-        },
-        state,
-      },
-    });
-    queue.sent = [];
-
-    await expect(
-      processConversationWork(CONVERSATION_ID, {
-        nowMs: () => currentNowMs,
-        queue,
-        state,
-        run: createSlackConversationWorker({
-          getSlackAdapter: () => slackAdapter,
-          runtime: {
-            handleNewMention: async () => {
-              currentNowMs = 242_000;
-            },
-            handleSubscribedMessage: async () => {
-              throw new Error("unexpected subscribed route");
-            },
-          },
-          state,
-        }),
-      }),
-    ).resolves.toEqual({ status: "completed" });
-
-    expect(queue.sent).toEqual([]);
-    const work = await getConversationWorkState({
-      conversationId: CONVERSATION_ID,
-      state,
-    });
-    expect(work?.needsRun).toBe(false);
-    expect(work ? countPendingConversationMessages(work) : 0).toBe(0);
   });
 
   it("rejects malformed Vercel Queue payloads", async () => {
